requirements.txt
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
msal==1.24.1
schedule==1.2.0
python-dateutil==2.8.2
